Address issue #1 where the plug returns no data
Add newline terminators to both the capture script and the Node-RED
flow.

Between plug firmware versions 4294 and 4918 there was a change
requiring the subscribe command to be terminated by a newline. The older
versions of the plug accept both newline terminated and non-terminated
commands, whereas the new version only accepts the former.